requirement.txt
asgiref==3.7.1
coverage==7.2.6
Django==4.2.1
djangorestframework==3.14.0
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
psycopg2==2.9.6
pytest==7.3.1
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-mock==3.10.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4